+++ src/main/java/com/myecommerce/MyECommerce/mapper/MemberMapper.java
@@ -1,6 +1,7 @@
 package com.myecommerce.MyECommerce.mapper;
 
-import com.myecommerce.MyECommerce.dto.MemberDto;
+import com.myecommerce.MyECommerce.dto.member.RequestMemberDto;
+import com.myecommerce.MyECommerce.dto.member.ResponseMemberDto;
 import com.myecommerce.MyECommerce.entity.member.Member;
 import org.mapstruct.Mapper;
 import org.mapstruct.factory.Mappers;
@@ -12,6 +13,6 @@ public interface MemberMapper {
 
     MemberMapper INSTANCE = Mappers.getMapper(MemberMapper.class);
 
-    MemberDto toDto(Member member);
-    Member toEntity(MemberDto memberDto);
+    ResponseMemberDto toDto(Member member);
+    Member toEntity(RequestMemberDto memberDto);
 }
